backend/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.10
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.1.0.post1
sqlalchemy>=2.0
asyncpg>=0.29
httpx>=0.27
cryptography>=42.0
python-multipart